v2 - Breaking change: no shadowRoot observed
While developing element-notifier I have realized that simply
crawling optional shadowRoot, if open and present, is not enough.

The element with it get notified in insertion, but if an inner element
gets removed, it will never pass through the disconnected dance.

As this has been broken for the whole v1 time, it is now needed,
as explained in https://github.com/WebReflection/element-notifier#about-shadowdom
that whoever needs to obseve properly ShadowDOM content, needs to explicitly
opt-in through the exposed observer per each QSAO instance.